Supplement live stream search with a "top streams" source.
This allieviates an annoying double-feature of live stream searching.
Now, the user is presented with "top live streams" when there is no
pattern to search with, and provides live stream search as a separate
source.

Top live streams are also cached, which means that `helm-twitch` should
now have near-instant re-opening (as long as the cache is < 60s old).

Signed-off-by: Aaron Jacobs <atheriel@gmail.com>